backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
asyncpg==0.30.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
pillow==10.3.0
python-dotenv==1.0.1
alembic==1.13.1
pydantic[email]==2.7.1
pydantic-settings==2.3.0
